Big update

+++ src/main/kotlin/fr/lewon/circuit/designer/model/geometry/Point.kt
@@ -1,3 +1,3 @@
 package fr.lewon.circuit.designer.model.geometry
 
-class Point(val x: Int, val y: Int)
+class Point(val x: Double, val y: Double)
